+++ src/client/logging/logger.ts
@@ -11,7 +11,6 @@ import { getConsoleTransport, getFileTransport, isConsoleTransport } from './tra
 import { Arguments } from './util';
 
 export type LoggerConfig = {
-    level?: LogLevel;
     file?: {
         logfile: string;
     };
@@ -38,13 +37,6 @@ interface IConfigurableLogger {
 
 // Set up a logger just the way we like it.
 export function configureLogger(logger: IConfigurableLogger, config: LoggerConfig) {
-    if (config.level) {
-        const levelName = resolveLevelName(config.level, winston.config.npm.levels);
-        if (levelName) {
-            logger.level = levelName;
-        }
-    }
-
     if (config.file) {
         const formatter = getFormatter();
         const transport = getFileTransport(config.file.logfile, formatter);
